\name{data}
\alias{data}
\docType{data}
\title{
Example data
}
\description{
Synthetic dataset
}
\usage{data("data")}
\format{
  Precinct-level election data from the 2011 Canadian parliamentary elections..
  \describe{
    \item{\code{N}}{number of registered voters}
    \item{\code{t}}{turnout}
    \item{\code{v}}{votes for the Conservative party}
  }
}
\examples{
data(data)
}
\keyword{datasets}
