\name{spikes-internal}
\alias{a}
\alias{init}
\alias{range01 }
\alias{samp.mat}
\alias{y.v}
\alias{roundr}
\alias{dmbetabinom}
\alias{dmbeta}
\alias{rmnorm}
\alias{resample}
\alias{em.update}
\alias{fraud}
\alias{bupdate}
\alias{est.density}
\alias{em.update}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal functions
}
\description{
Internal functions, should not be called by user
}
